+++ rspec-core/Changelog.md
@@ -35,6 +35,7 @@ Breaking Changes:
   (Phil Pirozhkov, rspec/rspec-core#2864)
 * Unify multi-condition filtering to use "all" semantic. (Phil Pirozhkov, rspec/rspec-core#2874)
 * Change the default order to random. (Santiago Bartesaghi, rspec/rspec-core#2929)
+* Raise an error when a block is passed to `include_examples`. (Phil Pirozhkov, rspec/rspec#259)
 
 Enhancements: